+++ src/main/java/com/booleanuk/core/Coffee.java
@@ -2,7 +2,7 @@ package com.booleanuk.core;
 
 public class Coffee extends Item{
 
-    public Coffee(String sku, Double price, String name, String variant){
+    public Coffee(String sku, Integer price, String name, String variant){
         super(sku, price, name, variant);
     }
 }
